chore: split MapViewOverlay to another file
